openssl 3.2.0
It's now possible to build on Tiger with asm support.
Disabling HWAES fixes build on Tiger & Leopard with asm.
OPENSSL_LOCAL_CONFIG_DIR landed in OpenSSL 1.1.0 and we've building
with it untouched in openssl.rb, don't bother unsetting it here too.
Skip setting compiler optimisation, the build ends up setting -O3 in
places.
With these changes + the removal of makedepend on build system results
in a build just short of 2 hours (116.8 minutes) on a G5.