fpc 3.2.2 / 3.0.4
Freepascal 3.2.2 needs a minimum version of 3.0.4 to bootstrap.
v3.2.2 needs 10.9 or newer to build on Intel darwin.
It tries to target 10.8 but fails.
It is not possible to build the Intel target compiler on PowerPC.
So build the latest version on PowerPC systems & those running >=10.9.
For 10.4 to 10.8 Intel builds stick with 3.0.4.